+++ NEWS.md
@@ -4,6 +4,7 @@ v0.2.4
 -------
 
 - Add `@assign`.
+- Add support for lazy compute functions that return `Base.Broadcast.Broadcasted` objects. 
 
 v0.2.3
 -------